pyproject.toml
[tool.poetry]
name = "rabbitmq-clients"
version = "0.1.19"
description = "High level interfaces for work with RabbitMQ."
authors = ["deskent <battenetciz@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = 'https://github.com/Deskent/rabbit_mq_client'

[tool.poetry.dependencies]
python = "^3.10"
aio-pika = "^9.4.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
ruff = "^0.4.10"
flake8 = "^7.1.0"
pytest-asyncio = "^0.23.7"
pre-commit = "^3.7.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "env",
    "migrations",
    "spam",
    "*.toml",
    "test_migrations/staff/utils",
]
line-length = 79

# Assume Python 3.10
target-version = "py310"


[tool.ruff.format]
## Like Black, use single quotes for strings.
quote-style = "single"

## Like Black, indent with spaces, rather than tabs.
indent-style = "space"

## Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

## Like Black, automatically detect the appropriate line ending.
line-ending = "auto"


[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true
lines_after_imports = 2
lexicographical = true
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
no_lines_before = ["STDLIB"]
known_third_party = []
known_local_folder = []
profile = "hug"
src_paths = ["src", "tests"]
skip_glob = ["migrations"]

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # The following are specific to Black, you probably don't want those.
  tests/data
  | profiling
)/
'''
# We use preview style for formatting Black itself. If you
# want stable formatting across releases, you should keep
# this off.
preview = false
skip-string-normalization = true
